+++ frontend/consumer-ui/components/Map/ShopCard.stories.js
