+++ frontend/consumer-ui/components/Map/ShopCard.vue
