+++ frontend/consumer-ui/pages/index.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-import Map from '~/components/Map'
+import Map from '~/components/Map/Map'
 
 export default {
   components: {
